Implementacion de dialog para confirmar eliminacion de usuario

+++ cede-angular/src/app/users/list/list.component.ts
@@ -2,8 +2,9 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { UsersService } from '../users.service';
 
 import { UserModel } from '../users.model';
-import { MatTableDataSource } from '@angular/material';
-
+import { MatTableDataSource, MatDialog } from '@angular/material';
+import { DialogComponent } from './../../shared/dialog/dialog.component'
+import { TranslateService } from '@ngx-translate/core';
 
 @Component({
   selector: 'app-list',
@@ -13,8 +14,8 @@ import { MatTableDataSource } from '@angular/material';
 export class ListComponent implements OnInit, OnDestroy {
   displayedColumns: string[] = ['userId', 'name', 'lastName', 'nit', 'birthDay', 'email', 'delete'];
   dataSource: MatTableDataSource<UserModel>;
-  isLoading:boolean=true;
-  constructor(private usersService: UsersService) {
+  isLoading: boolean = true;
+  constructor(private usersService: UsersService, private dialog: MatDialog, private translate: TranslateService) {
 
   }
 
@@ -30,17 +31,36 @@ export class ListComponent implements OnInit, OnDestroy {
   load() {
     this.usersService.getAll().subscribe(
       (data: any) => {
-        this.isLoading=false;
+        this.isLoading = false;
         this.dataSource = new MatTableDataSource(data);
       },
       (error: any) => {
-        this.isLoading=false;
+        this.isLoading = false;
       });
   }
 
-  delete(element){
-    this.usersService.delete(element.userId).subscribe((data:any)=>{
-        this.load();
+  delete(element) {
+    this.isLoading = true;
+
+    this.translate.get("users.confirmMessage").subscribe((message: string) => {
+      let dialogRef = this.dialog.open(DialogComponent, {
+        data: {
+          title: element.name,
+          message: message
+        }
+      })
+      dialogRef.afterClosed().subscribe(result => {
+        if (result) {
+          this.usersService.delete(element.userId).subscribe((data: any) => {
+            this.isLoading = false;
+            this.load();
+          });
+        } else {
+          this.isLoading = false;
+        }
+      });
     });
+
+
   }
 }
